+++ crates/sdk-core/tests/heavy_tests.rs
@@ -84,8 +84,12 @@ async fn activity_load() {
     let mut starter = CoreWfStarter::new("activity_load");
     starter.sdk_config.max_cached_workflows = CONCURRENCY;
     starter.sdk_config.activity_task_poller_behavior = Some(PollerBehavior::SimpleMaximum(10));
-    starter.sdk_config.tuner =
-        Arc::new(TunerHolder::fixed_size(CONCURRENCY, CONCURRENCY, 100, 100));
+    starter.set_core_tuner(Arc::new(TunerHolder::fixed_size(
+        CONCURRENCY,
+        CONCURRENCY,
+        100,
+        100,
+    )));
     starter.sdk_config.register_activities(StdActivities);
     starter
         .sdk_config
@@ -177,7 +181,7 @@ async fn chunky_activities_resource_based() {
             Duration::from_millis(0),
         ))
         .with_activity_slots_options(ResourceSlotOptions::new(5, 1000, Duration::from_millis(50)));
-    starter.sdk_config.tuner = Arc::new(tuner);
+    starter.set_core_tuner(Arc::new(tuner));
 
     starter.sdk_config.register_activities(ChunkyActivities);
     starter
@@ -254,7 +258,7 @@ async fn workflow_load() {
     let mut starter = CoreWfStarter::new_with_runtime("workflow_load", rt);
     starter.sdk_config.max_cached_workflows = 200;
     starter.sdk_config.activity_task_poller_behavior = Some(PollerBehavior::SimpleMaximum(10));
-    starter.sdk_config.tuner = Arc::new(TunerHolder::fixed_size(5, 100, 100, 100));
+    starter.set_core_tuner(Arc::new(TunerHolder::fixed_size(5, 100, 100, 100)));
     starter.sdk_config.register_activities(StdActivities);
     let task_queue = starter.get_task_queue().to_owned();
     starter
@@ -310,7 +314,7 @@ async fn evict_while_la_running_no_interference() {
     // Though it doesn't make sense to set wft higher than cached workflows, leaving this commented
     // introduces more instability that can be useful in the test.
     // starter.max_wft(20);
-    starter.sdk_config.tuner = Arc::new(TunerHolder::fixed_size(100, 10, 20, 1));
+    starter.set_core_tuner(Arc::new(TunerHolder::fixed_size(100, 10, 20, 1)));
     starter.sdk_config.register_activities(StdActivities);
     starter
         .sdk_config
@@ -466,7 +470,7 @@ async fn poller_autoscaling_basic_loadtest() {
     let wf_name = "poller_load";
     let mut starter = CoreWfStarter::new("poller_load");
     starter.sdk_config.max_cached_workflows = 5000;
-    starter.sdk_config.tuner = Arc::new(TunerHolder::fixed_size(1000, 1000, 100, 1));
+    starter.set_core_tuner(Arc::new(TunerHolder::fixed_size(1000, 1000, 100, 1)));
     starter.sdk_config.workflow_task_poller_behavior = Some(PollerBehavior::Autoscaling(
         AutoscalingOptions::builder()
             .minimum(1)