+++ crates/sdk-core/tests/integ_tests/workflow_tests.rs
@@ -389,7 +389,7 @@ async fn wft_timeout_doesnt_create_unsolvable_autocomplete() {
     let mut wf_starter = CoreWfStarter::new("wft_timeout_doesnt_create_unsolvable_autocomplete");
     // Test needs eviction on and a short timeout
     wf_starter.sdk_config.max_cached_workflows = 0_usize;
-    wf_starter.sdk_config.tuner = Arc::new(TunerHolder::fixed_size(1, 1, 1, 1));
+    wf_starter.set_core_tuner(Arc::new(TunerHolder::fixed_size(1, 1, 1, 1)));
     wf_starter.sdk_config.workflow_task_poller_behavior =
         Some(PollerBehavior::SimpleMaximum(1_usize));
     wf_starter.workflow_options.task_timeout = Some(Duration::from_secs(1));
@@ -518,7 +518,7 @@ impl SlowCompletesWf {
 async fn slow_completes_with_small_cache() {
     let wf_name = "slow_completes_with_small_cache";
     let mut starter = CoreWfStarter::new(wf_name);
-    starter.sdk_config.tuner = Arc::new(TunerHolder::fixed_size(5, 10, 1, 1));
+    starter.set_core_tuner(Arc::new(TunerHolder::fixed_size(5, 10, 1, 1)));
     starter.sdk_config.max_cached_workflows = 5_usize;
     starter.sdk_config.register_activities(StdActivities);
     starter